[dagit] Move Code location and Daemon status polling to context
