+++ js_modules/dagit/packages/core/src/instance/DeploymentStatusType.tsx
@@ -1,0 +1,4 @@
+export type StatusAndMessage = {
+  type: 'warning' | 'spinner';
+  content: string | JSX.Element;
+};
