+++ js_modules/dagit/packages/core/src/instance/InstanceTabs.tsx
@@ -4,6 +4,8 @@ import * as React from 'react';
 
 import {useFeatureFlags} from '../app/Flags';
 import {QueryRefreshCountdown, QueryRefreshState} from '../app/QueryRefresh';
+import {InstanceWarningIcon} from '../nav/InstanceWarningIcon';
+import {WorkspaceStatus} from '../nav/WorkspaceStatus';
 import {TabLink} from '../ui/TabLink';
 
 import {InstancePageContext} from './InstancePageContext';
@@ -29,9 +31,19 @@ export const InstanceTabs = <TData extends Record<string, any>>(props: Props<TDa
           <TabLink id="overview" title="Overview" to="/instance/overview" />
         )}
         {flagNewWorkspace ? (
-          <TabLink id="code-locations" title="Code locations" to="/instance/code-locations" />
+          <TabLink
+            id="code-locations"
+            title="Code locations"
+            to="/instance/code-locations"
+            icon={<WorkspaceStatus placeholder={false} />}
+          />
         ) : null}
-        <TabLink id="health" title={healthTitle} to="/instance/health" />
+        <TabLink
+          id="health"
+          title={healthTitle}
+          to="/instance/health"
+          icon={<InstanceWarningIcon />}
+        />
         {flagNewWorkspace ? null : (
           <TabLink id="schedules" title="Schedules" to="/instance/schedules" />
         )}
